+++ CHANGELOG.md
@@ -12,7 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Added CHANGELOG files
-- Added [World Scale Object](https://m-a.nadena.dev/docs/reference/world-scale-object)
+- [#1482] Added support for replacing pre-existing animator controllers to `Merge Animator`
+- [#1481] Added [World Scale Object](https://m-a.nadena.dev/docs/reference/world-scale-object)
 
 ### Fixed
 - [#1460] When importing parameter assets in MA Parameters, "local only" parameters were incorrectly treated as